refactor(listener): extract control input handlers
