+++ src/websocket/listener/conversation-runtime.ts
@@ -1,0 +1,74 @@
+import { QueueRuntime } from "../../queue/queueRuntime";
+import { scheduleQueueEmit } from "./protocol-outbound";
+import { getQueueItemScope, getQueueItemsScope } from "./queue";
+import {
+  evictConversationRuntimeIfIdle,
+  getOrCreateConversationRuntime,
+} from "./runtime";
+import type { ConversationRuntime, ListenerRuntime } from "./types";
+
+export function ensureConversationQueueRuntime(
+  listener: ListenerRuntime,
+  runtime: ConversationRuntime,
+): ConversationRuntime {
+  if (runtime.queueRuntime) {
+    return runtime;
+  }
+  runtime.queueRuntime = new QueueRuntime({
+    callbacks: {
+      onEnqueued: (item, queueLen) => {
+        runtime.pendingTurns = queueLen;
+        scheduleQueueEmit(listener, getQueueItemScope(item));
+      },
+      onDequeued: (batch) => {
+        runtime.pendingTurns = batch.queueLenAfter;
+        scheduleQueueEmit(listener, getQueueItemsScope(batch.items));
+      },
+      onBlocked: () => {
+        scheduleQueueEmit(listener, {
+          agent_id: runtime.agentId,
+          conversation_id: runtime.conversationId,
+        });
+      },
+      onCleared: (_reason, _clearedCount, items) => {
+        runtime.pendingTurns = 0;
+        scheduleQueueEmit(listener, getQueueItemsScope(items));
+        evictConversationRuntimeIfIdle(runtime);
+      },
+      onDropped: (item, _reason, queueLen) => {
+        runtime.pendingTurns = queueLen;
+        runtime.queuedMessagesByItemId.delete(item.id);
+        scheduleQueueEmit(listener, getQueueItemScope(item));
+        evictConversationRuntimeIfIdle(runtime);
+      },
+    },
+  });
+  return runtime;
+}
+
+export function getOrCreateScopedRuntime(
+  listener: ListenerRuntime,
+  agentId?: string | null,
+  conversationId?: string | null,
+): ConversationRuntime {
+  return ensureConversationQueueRuntime(
+    listener,
+    getOrCreateConversationRuntime(listener, agentId, conversationId),
+  );
+}
+
+/**
+ * Fallback for unscoped task notifications (e.g., reflection/init spawned
+ * outside turn processing). Picks the first ConversationRuntime that has a
+ * QueueRuntime, or null if none exist.
+ */
+export function findFallbackRuntime(
+  listener: ListenerRuntime,
+): ConversationRuntime | null {
+  for (const cr of listener.conversationRuntimes.values()) {
+    if (cr.queueRuntime) {
+      return cr;
+    }
+  }
+  return null;
+}
